requirements.txt
anthropic>=0.40.0
playwright>=1.44.0
beautifulsoup4>=4.12.0
requests>=2.31.0
streamlit>=1.35.0
pandas>=2.2.0
pyyaml>=6.0
python-dotenv>=1.0.0
mcp>=1.0.0
pdfplumber>=0.10.0
python-docx>=1.1.0
